% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{action_2}
\alias{action_2}
\title{action_2}
\usage{
action_2(x)
}
\arguments{
\item{x}{a list whose structure should be checked and cannot be simplified}
}
\value{
a list containing cleaned data
}
\description{
action_2
}
