% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{action_3}
\alias{action_3}
\title{action_3}
\usage{
action_3(x)
}
\arguments{
\item{x}{a list containing tibble, which should simplified and transform into character}
}
\value{
a character vector with the cleaned data
}
\description{
action_3
}
