% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_scrapping.R
\name{add_authors}
\alias{add_authors}
\title{add_authors}
\usage{
add_authors(ids, n_batchs, batch_size, path = NULL)
}
\arguments{
\item{ids}{A tibble containing a column dc_identifier}

\item{n_batchs}{Corresponds to the number of batch to scrape}

\item{batch_size}{Corresponds to the size of the batchs (max25)}

\item{path}{Path ot the current working directory}
}
\value{
A sample of the dataset which has been scrapped
}
\description{
This function automatially update the current dataset authors by geting new information on not scrapped authors
}
