% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_scrapping.R
\name{add_publications_authors}
\alias{add_publications_authors}
\title{add_publications_authors}
\usage{
add_publications_authors(data, n_authors, path = NULL, shuffle = T)
}
\arguments{
\item{data}{A tibble containing a column dc_identifier}

\item{n_authors}{Number of authors whose publications have to be scraped}

\item{path}{Path ot the current working directory}

\item{shuffle}{Logical value, if T then order of scrape is random}
}
\value{
A sample of the dataset which has been scrapped
}
\description{
This function automatially update the current dataset "publications_authors" by geting new information on non scrapped authors
}
