% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_cleaner.R
\name{clean_affiliations}
\alias{clean_affiliations}
\title{clean_affiliations}
\usage{
clean_affiliations(path, n = 1000, load_main = T)
}
\arguments{
\item{path}{Path of the working directory}

\item{n}{Number of authors whose affiliations should be parsed}
}
\value{
the updated version of affiliations_final
}
\description{
This function updates the affiliations file in regard of updated authors
}
