% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_cleaner.R
\name{clean_authors}
\alias{clean_authors}
\title{clean_authors}
\usage{
clean_authors(step1)
}
\arguments{
\item{step1}{a tibble containing authors to clean (response from parse_authors)}
}
\value{
the cleaned tibble
}
\description{
clean_authors
}
