% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_cleaner.R
\name{clean_authors_old}
\alias{clean_authors_old}
\title{clean_authors_old}
\usage{
clean_authors_old(authors, n, path, loading_main)
}
\arguments{
\item{authors}{a tibble containing authors to clean (response from authors scrapping)}
}
\value{
the cleaned tibble
}
\description{
clean_authors_old
}
