% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_cleaner.R
\name{clean_publications}
\alias{clean_publications}
\title{clean_publications}
\usage{
clean_publications(publications, n, path, loading_main = T)
}
\arguments{
\item{n}{Number of authors whose affiliations should be parsed}

\item{path}{Path of the working directory}
}
\value{
the updated version of affiliations_final
}
\description{
This function parses new scrapped publications and update the publications_final file
}
