% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_publication.R
\name{entry_core}
\alias{entry_core}
\title{gentry_core}
\usage{
entry_core(.x)
}
\arguments{
\item{.x}{a tibble with a dc_identifier column indicating the ids to scrape}
}
\value{
entries, is a list containing the publications of each scrapped author id
}
\description{
gentry_core
}
