% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funs.R
\name{funs}
\alias{funs}
\title{Create a list of functions calls.}
\usage{
funs(..., .args = list())
}
\arguments{
\item{...}{A list of functions specified by:

 - Their name, `"mean"`
 - The function itself, `mean`
 - A call to the function with `.` as a dummy argument,
   `mean(., na.rm = TRUE)`

 These arguments are automatically [quoted][rlang::quo]. They
 support [unquoting][rlang::quasiquotation] and splicing. See
 `vignette("programming")` for an introduction to these concepts.

 The following notations are **not** supported, see examples:

 - An anonymous function, `function(x) mean(x, na.rm = TRUE)`
 - An anonymous function in \pkg{purrr} notation, `~mean(., na.rm = TRUE)`}

\item{.args, args}{A named list of additional arguments to be added
to all function calls.}
}
\description{
\badgesoftdeprecated
}
\details{
`funs()` provides a flexible way to generate a named list of
functions for input to other functions like [summarise_at()].
}
\examples{
funs(mean, "mean", mean(., na.rm = TRUE))

# Override default names
funs(m1 = mean, m2 = "mean", m3 = mean(., na.rm = TRUE))

# If you have function names in a vector, use funs_
fs <- c("min", "max")
funs_(fs)

# Not supported
\dontrun{
funs(function(x) mean(x, na.rm = TRUE))
funs(~mean(x, na.rm = TRUE))}
}
