% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_scrapping.R
\name{get_publication_authors}
\alias{get_publication_authors}
\title{get_publication_authors}
\usage{
get_publication_authors(data, n_authors, path = NULL, shuffle = T)
}
\arguments{
\item{data}{A tibble containing a column dc_identifier}

\item{n_authors}{Number of authors whose publications have to be scraped}

\item{path}{Path ot the current working directory}

\item{shuffle}{Logical value, if T then order of scrape is random}
}
\value{
Raw Scrapped Informations
}
\description{
get_publication_authors
}
