% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{list_to_tibble}
\alias{list_to_tibble}
\title{listify}
\usage{
list_to_tibble(x)
}
\arguments{
\item{x}{a list to be transformed into a tibble}
}
\value{
a tibble with unique and cleaned column names
}
\description{
This function check whether the input is a list on which we could map. If this is not the case, it encapsualit the input in a list.
}
