% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{list_unsimplifiable}
\alias{list_unsimplifiable}
\title{list_simplifiable}
\usage{
list_unsimplifiable(x)
}
\arguments{
\item{x}{a variable to check the type for the parsing process}
}
\value{
a logical vector T if the variable is a list and cannot be simplifiable
}
\description{
list_simplifiable
}
