% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{map_select}
\alias{map_select}
\title{map_select}
\usage{
map_select(x, vars)
}
\arguments{
\item{x}{a list from which columns should be selected}

\item{vars}{a vector of names to select in x}
}
\value{
the selected columns (if not contained, replaced by NA)
}
\description{
map_select
}
