% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{mutate_pos}
\alias{mutate_pos}
\title{mutate_pos}
\usage{
mutate_pos(data, ...)
}
\arguments{
\item{data}{a dataframe with var to mutate}

\item{...}{arguments to be aplly to mutate}
}
\value{
mutated dataframe
}
\description{
mutate_pos
}
