% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_cleaner.R
\name{parse_authors}
\alias{parse_authors}
\title{parse_authors}
\usage{
parse_authors(authors, n)
}
\arguments{
\item{authors}{a tibble containing authors to parse}
}
\value{
parsed authors
}
\description{
parse_authors
}
