% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_scrapping.R
\name{parse_publication_authors}
\alias{parse_publication_authors}
\title{parse_publication_authors}
\usage{
parse_publication_authors(step1)
}
\arguments{
\item{step1}{Logical value, if T then order of scrape is random}

\item{path}{Path ot the current working directory}
}
\value{
Parsed Scrapped Informations
}
\description{
parse_publication_authors
}
