% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{progressively}
\alias{progressively}
\title{progressively}
\usage{
progressively(.f, .n, ...)
}
\arguments{
\item{.f}{a function to make progressive}

\item{.n}{input length}

\item{...}{further parameters for the input function}
}
\value{
the same function but progressively
}
\description{
progressively
}
