% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{rename_if_included}
\alias{rename_if_included}
\title{rename_if_included}
\usage{
rename_if_included(data, x, y)
}
\arguments{
\item{data}{a tibble}

\item{x}{a string as a variable name to be checked for}

\item{y}{a string replacing x, if x machtes}
}
\value{
a tibble with x renamed y
}
\description{
rename_if_included
}
