% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{rename_pos}
\alias{rename_pos}
\title{rename_pos}
\usage{
rename_pos(data, ...)
}
\arguments{
\item{data}{a dataframe with var to rename}

\item{...}{arguments to be aplly to rename}
}
\value{
dataframe with renamed columns
}
\description{
rename_pos
}
