% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_cleaner.R
\name{scopus_clean_author_publications}
\alias{scopus_clean_author_publications}
\title{scopus_clean_author_publications}
\usage{
scopus_clean_author_publications(author_publications)
}
\arguments{
\item{author_publications}{a tibble containing author_publications to clean (response from author_publications scrapping)}
}
\value{
the cleaned tibble
}
\description{
scopus_clean_author_publications
}
