% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_cleaner.R
\name{scopus_clean_authors}
\alias{scopus_clean_authors}
\title{scopus_clean_authors}
\usage{
scopus_clean_authors(authors)
}
\arguments{
\item{authors}{a tibble containing authors to clean (response from authors scrapping)}
}
\value{
the cleaned tibble
}
\description{
scopus_clean_authors
}
