% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_cleaner.R
\name{scopus_clean_citations}
\alias{scopus_clean_citations}
\title{scopus_clean_citations}
\usage{
scopus_clean_citations(publications)
}
\arguments{
\item{publications}{a tibble containing publications from which citation matrix should be extracted (response from publications scrapping)}
}
\value{
the cleaned citation matrix as tibble with citing and cited ids
}
\description{
scopus_clean_citations
}
