% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_cleaner.R
\name{scopus_clean_publications}
\alias{scopus_clean_publications}
\title{scopus_clean_publications}
\usage{
scopus_clean_publications(publications)
}
\arguments{
\item{publications}{a tibble containing publications to clean (response from publications scrapping)}
}
\value{
the cleaned tibble
}
\description{
scopus_clean_publications
}
