% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{select_pos}
\alias{select_pos}
\title{select_pos}
\usage{
select_pos(data, vars)
}
\arguments{
\item{data}{a dataframe with columns to select}

\item{vars}{a vector of names to select in x}
}
\value{
selected columns
}
\description{
select_pos
}
