% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{simplifiable}
\alias{simplifiable}
\title{simplifiable}
\usage{
simplifiable(x)
}
\arguments{
\item{x}{a variable to map on}
}
\value{
a logical vector T if the variable can be simplified
}
\description{
The variable check whether a list element can be simplified
}
