% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{simplify_vector}
\alias{simplify_vector}
\title{simplify_vector}
\usage{
simplify_vector(x)
}
\arguments{
\item{x}{a variable with tibble to simplify}
}
\value{
a character vector
}
\description{
Use to simplify variables whose elements are unnecessarily nested into tibble. It just get the value embedded in the tibble and return a character vector.
}
