% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{transform_list}
\alias{transform_list}
\title{transform_list}
\usage{
transform_list(x, nam = "x")
}
\arguments{
\item{x}{a response to parse}
}
\value{
a tibble containing parsed respone
}
\description{
transform_list
}
