% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{transform_list_old}
\alias{transform_list_old}
\title{transform_list_old}
\usage{
transform_list_old(x)
}
\arguments{
\item{x}{a response to parse}
}
\value{
a tibble containing parsed respone
}
\description{
transform_list_old
}
