% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{unlist_tibble_core}
\alias{unlist_tibble_core}
\title{unlist_tibble_core}
\usage{
unlist_tibble_core(.x)
}
\arguments{
\item{.x}{a list to transform into a tibble}
}
\value{
a tibble whose structure matchs the input structure
}
\description{
unlist_tibble_core
}
