% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{unnest_pos}
\alias{unnest_pos}
\alias{unnest_pos}
\title{unnest_pos}
\usage{
unnest_pos(data, ...)

unnest_pos(data, ...)
}
\arguments{
\item{data}{a dataframe with var to unnest}

\item{...}{arguments to be aplly to unnest}
}
\value{
dataframe with unnested columns
}
\description{
unnest_pos

unnest_pos
}
