% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{unnest_tibble}
\alias{unnest_tibble}
\title{unnest_tibble}
\usage{
unnest_tibble(x)
}
\arguments{
\item{x}{a tibble whose structure should be cleaned}
}
\value{
a tibble with cleaned structure
}
\description{
This function check the internal structure of an element an clean it
}
