% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scopus_helper.R
\name{unselect_pos}
\alias{unselect_pos}
\title{unselect_pos}
\usage{
unselect_pos(data, vars)
}
\arguments{
\item{data}{a dataframe with columns to unselect}

\item{vars}{a vector of names to unselect in x}
}
\value{
unselected columns
}
\description{
unselect_pos
}
