% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_cleaner.R
\name{update_citation_mat}
\alias{update_citation_mat}
\title{update_citation_mat}
\usage{
update_citation_mat()
}
\value{
the cleaned citation matrix as tibble with citing and cited ids
}
\description{
This function takes the updated publications_final and extract the new citation matrix
}
