% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_scrapping.R
\name{add_abstracts}
\alias{add_abstracts}
\title{add_abstracts}
\usage{
add_abstracts(data, n_abstracts, path = NULL)
}
\arguments{
\item{data}{A tibble containing a column dc_identifier}

\item{n_abstracts}{Number of abstracts to scrape}

\item{path}{Path ot the current working directory}
}
\value{
A sample of the dataset which has been scrapped
}
\description{
This function automatially update the current dataset "publications" by geting new information on non scrapped entries
}
