% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_cleaner.R
\name{clean_authors}
\alias{clean_authors}
\title{clean_authors}
\usage{
clean_authors(authors, n, path, loading_main)
}
\arguments{
\item{authors}{a tibble containing authors to clean (response from authors scrapping)}
}
\value{
the cleaned tibble
}
\description{
clean_authors
}
